Capture link for case detail page
Updated parser to capture the link for the case detail page. Removed
parentheses to conform to style guide. Added a wrapper class so that the
first two records are auto deleted for when parser is called. Moved
classes to seperate files.

This is done so that the case detail page can be parsed automatically.